requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
openai==1.30.1
pypdf2==3.0.1
faiss-cpu==1.13.0
python-dotenv==1.0.1
tiktoken==0.7.0